+++ src/common/Track.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const Track = (props) => {
+const Track = props => {
   const { className, included, vertical, style } = props;
   let { length, offset, reverse } = props;
   if (length < 0) {
@@ -9,15 +9,17 @@ const Track = (props) => {
     offset = 100 - offset;
   }
 
-  const positonStyle = vertical ? {
-    [reverse ? 'top' : 'bottom']: `${offset}%`,
-    [reverse ? 'bottom' : 'top']: 'auto',
-    height: `${length}%`,
-  } : {
-    [reverse ? 'right' : 'left']: `${offset}%`,
-    [reverse ? 'left' : 'right']: 'auto',
-    width: `${length}%`,
-  };
+  const positonStyle = vertical
+    ? {
+        [reverse ? 'top' : 'bottom']: `${offset}%`,
+        [reverse ? 'bottom' : 'top']: 'auto',
+        height: `${length}%`,
+      }
+    : {
+        [reverse ? 'right' : 'left']: `${offset}%`,
+        [reverse ? 'left' : 'right']: 'auto',
+        width: `${length}%`,
+      };
 
   const elStyle = {
     ...style,
