+++ src/common/createSlider.tsx
@@ -52,9 +52,7 @@ export default function createSlider(Component) {
       const isPointDiffEven = isFinite(max - min) ? (max - min) % step === 0 : true; // eslint-disable-line
       warning(
         step && Math.floor(step) === step ? isPointDiffEven : true,
-        'Slider[max] - Slider[min] (%s) should be a multiple of Slider[step] (%s)',
-        max - min,
-        step,
+        `Slider[max] - Slider[min] (${max - min}) should be a multiple of Slider[step] (${step})`,
       );
       this.handlesRefs = {};
     }
